src/api/proxy_test.go
package api

import (
	"net/http"
	"net/http/httptest"
	"testing"
)

func TestProxyGet(t *testing.T) {
	req, err := http.NewRequest("GET", "/proxy/", nil)
	if err != nil {
		t.Fatal(err)
	}
	rr := httptest.NewRecorder()
	handler := http.HandlerFunc(proxyGet)
	handler.ServeHTTP(rr, req)
	if status := rr.Code; status != http.StatusOK {
		t.Errorf("handler returned wrong status code: got %v want %v",
			status, http.StatusOK)
	}
}

func TestProxyPOST(t *testing.T) {
	req, err := http.NewRequest("POST", "/proxy/", nil)
	if err != nil {
		t.Fatal(err)
	}
	rr := httptest.NewRecorder()
	handler := http.HandlerFunc(proxyPost)
	handler.ServeHTTP(rr, req)
	if status := rr.Code; status != http.StatusOK {
		t.Errorf("handler returned wrong status code: got %v want %v",
			status, http.StatusOK)
	}
}
